+++ packages/ohm-js/src/pexprs-substituteParams.js
@@ -5,7 +5,7 @@
 // --------------------------------------------------------------------
 
 const common = require('./common');
-const pexprs = require('./pexprs');
+const pexprs = require('./pexprs-main');
 
 // --------------------------------------------------------------------
 // Operations
